// generated/models/ApiCollectOfferPriceReference.ts
export enum ApiCollectOfferPriceReferenceKindEnum {
    Bid = 'bid',
    Ask = 'ask'
}